deleted an unneeded js file. made some new node classes for use in the game, such as node, gridlockednode, and powerupnode. Did a more robust job on the powerup system that is super flexible and cool. flags will be coming soon. players now check to see if they collide with any of the items in the game state, if so and they don't have an item yet, they pick them up.

+++ demo/src/main/java/com/cycapservers/game/Node.java
@@ -2,32 +2,32 @@ package com.cycapservers.game;
 
 public class Node {
 	
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	public Node(Node n) {
 		this.x = n.x;
 		this.y = n.y;
 	}
 	
-	public Node(double x, double y) {
+	public Node(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 }
